Module TuringMachine__Run

val step : ('symbol, 'state) TuringMachine.Data.turingMachineID -> ('symbol, 'state) TuringMachine.Data.turingMachine -> ('symbol, 'state) TuringMachine.Data.turingMachineID option

This function aim to represent the behavior a of given machine. So, it give each step of the machine.

parameter id

This parameter is the instantaneous description of a machine.

parameter machine

This parameter is the machine which the id represent.

returns

The return is the next ID of the machine, this ID is given by the thansition function. Besides, the returned value is inside a option because when the given machine-ID is in a halt state, it return None .